#481 moved implementation after review.
