+++ logic/atrium-logic-jvm/src/main/kotlin/ch/tutteli/atrium/logic/impl/DefaultChronoLocalDateTimeAssertions.kt
@@ -5,6 +5,8 @@ import ch.tutteli.atrium.creating.AssertionContainer
 import ch.tutteli.atrium.logic.ChronoLocalDateTimeAssertions
 import ch.tutteli.atrium.logic.createDescriptiveAssertion
 import ch.tutteli.atrium.translations.DescriptionDateTimeLikeAssertion.*
+import java.time.LocalDate
+import java.time.LocalDateTime
 import java.time.chrono.ChronoLocalDate
 import java.time.chrono.ChronoLocalDateTime
 
@@ -14,6 +16,13 @@ class DefaultChronoLocalDateTimeAssertions : ChronoLocalDateTimeAssertions {
         expected: ChronoLocalDateTime<*>
     ): Assertion = container.createDescriptiveAssertion(IS_BEFORE, expected) { it.isBefore(expected) }
 
+    override fun <T : ChronoLocalDateTime<out ChronoLocalDate>> isBefore(
+        container: AssertionContainer<T>,
+        expected: String
+    ): Assertion = container.createDescriptiveAssertion(IS_BEFORE, expected) {
+        it.isBefore(stringToLocalDateTime(expected))
+    }
+
     override fun <T : ChronoLocalDateTime<out ChronoLocalDate>> isBeforeOrEqual(
         container: AssertionContainer<T>,
         expected: ChronoLocalDateTime<*>
@@ -21,10 +30,24 @@ class DefaultChronoLocalDateTimeAssertions : ChronoLocalDateTimeAssertions {
         it.isBefore(expected) || it.isEqual(expected)
     }
 
+    override fun <T : ChronoLocalDateTime<out ChronoLocalDate>> isBeforeOrEqual(
+        container: AssertionContainer<T>,
+        expected: String
+    ): Assertion = container.createDescriptiveAssertion(IS_BEFORE_OR_EQUAL, expected) {
+        it.isBefore(stringToLocalDateTime(expected)) || it.isEqual(stringToLocalDateTime(expected))
+    }
+
     override fun <T : ChronoLocalDateTime<out ChronoLocalDate>> isAfter(
         container: AssertionContainer<T>,
         expected: ChronoLocalDateTime<*>
     ): Assertion = container.createDescriptiveAssertion(IS_AFTER, expected) { it.isAfter(expected) }
+
+    override fun <T : ChronoLocalDateTime<out ChronoLocalDate>> isAfter(
+        container: AssertionContainer<T>,
+        expected: String
+    ): Assertion = container.createDescriptiveAssertion(IS_AFTER, expected) {
+        it.isAfter(stringToLocalDateTime(expected))
+    }
 
     override fun <T : ChronoLocalDateTime<out ChronoLocalDate>> isAfterOrEqual(
         container: AssertionContainer<T>,
@@ -33,10 +56,32 @@ class DefaultChronoLocalDateTimeAssertions : ChronoLocalDateTimeAssertions {
         it.isAfter(expected) || it.isEqual(expected)
     }
 
+    override fun <T : ChronoLocalDateTime<out ChronoLocalDate>> isAfterOrEqual(
+        container: AssertionContainer<T>,
+        expected: String
+    ): Assertion = container.createDescriptiveAssertion(IS_AFTER_OR_EQUAL, expected) {
+        it.isAfter(stringToLocalDateTime(expected)) || it.isEqual(stringToLocalDateTime(expected))
+    }
+
     override fun <T : ChronoLocalDateTime<out ChronoLocalDate>> isEqual(
         container: AssertionContainer<T>,
         expected: ChronoLocalDateTime<*>
     ): Assertion = container.createDescriptiveAssertion(IS_EQUAL_TO, expected) {
         it.isEqual(expected)
+    }
+
+    override fun <T : ChronoLocalDateTime<out ChronoLocalDate>> isEqual(
+        container: AssertionContainer<T>,
+        expected: String
+    ): Assertion = container.createDescriptiveAssertion(IS_EQUAL_TO, expected) {
+        it.isEqual(stringToLocalDateTime(expected))
+    }
+
+    private fun stringToLocalDateTime(data: String): LocalDateTime {
+        return if (data.contains("T")) {
+            LocalDateTime.parse(data)
+        } else {
+            LocalDate.parse(data).atStartOfDay()
+        }
     }
 }
