+++ misc/deprecated/domain/api/atrium-domain-api-jvm/src/main/kotlin/ch/tutteli/atrium/domain/creating/ChronoLocalDateTimeAssertions.kt
@@ -25,19 +25,9 @@ interface ChronoLocalDateTimeAssertions {
         expected: ChronoLocalDateTime<*>
     ): Assertion
 
-    fun <T : ChronoLocalDateTime<out ChronoLocalDate>> isBefore(
-        expect: Expect<T>,
-        expected: String
-    ): Assertion
-
     fun <T : ChronoLocalDateTime<out ChronoLocalDate>> isBeforeOrEquals(
         expect: Expect<T>,
         expected: ChronoLocalDateTime<*>
-    ): Assertion
-
-    fun <T : ChronoLocalDateTime<out ChronoLocalDate>> isBeforeOrEquals(
-        expect: Expect<T>,
-        expected: String
     ): Assertion
 
     fun <T : ChronoLocalDateTime<out ChronoLocalDate>> isAfter(
@@ -45,28 +35,13 @@ interface ChronoLocalDateTimeAssertions {
         expected: ChronoLocalDateTime<*>
     ): Assertion
 
-    fun <T : ChronoLocalDateTime<out ChronoLocalDate>> isAfter(
-        expect: Expect<T>,
-        expected: String
-    ): Assertion
-
     fun <T : ChronoLocalDateTime<out ChronoLocalDate>> isAfterOrEquals(
-        expect: Expect<T>,
-        expected: ChronoLocalDateTime<*>
-    ): Assertion
-
-    fun <T : ChronoLocalDateTime<out ChronoLocalDate>> isAfterOrEquals(
-        expect: Expect<T>,
-        expected: String
-    ): Assertion
-
-    fun <T : ChronoLocalDateTime<out ChronoLocalDate>> isEqual(
         expect: Expect<T>,
         expected: ChronoLocalDateTime<*>
     ): Assertion
 
     fun <T : ChronoLocalDateTime<out ChronoLocalDate>> isEqual(
         expect: Expect<T>,
-        expected: String
+        expected: ChronoLocalDateTime<*>
     ): Assertion
 }
